Add option only_one_perimeter_on_top
You can now set the option, to generate only one perimeter on surfaces with top infill.
You have the option to enable this feature for all surfaces with top infill or just for the topmost surface.
The option can be found in the advanced section of the perimeter settings and is available on the expert or advanced level.